Changes made to add comments to the files
